requirements.txt
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.0
Pillow==9.3.0
plotly==5.11.0
requests==2.28.1
streamlit==1.15.2
